+++ src/main/java/ru/spbau/mit/characters/Creep.java
@@ -1,6 +1,4 @@
-package ru.spbau.mit.mobs;
-
-import ru.spbau.mit.core.Stats;
+package ru.spbau.mit.characters;
 
 public class Creep extends Mob {
     private static final char DEFAULT_ICON = '♞';
