+++ src/main/java/ru/spbau/mit/characters/Mob.java
@@ -1,7 +1,4 @@
-package ru.spbau.mit.mobs;
-
-import ru.spbau.mit.core.Character;
-import ru.spbau.mit.core.Stats;
+package ru.spbau.mit.characters;
 
 public abstract class Mob extends Character {
     public Mob(char icon, Stats baseStats) {
